+++ src/meer21cm/fg.py
@@ -1,12 +1,9 @@
 import healpy as hp
 import numpy as np
 import meer21cm
-from meer21cm.util import read_healpix_fits
-from meer21cm.util import convert_hpmap_in_jy_to_temp
-from meer21cm.util import healpix_to_wcs
+from meer21cm.util import healpix_to_wcs, read_healpix_fits, check_unit_equiv
 from astropy import units
-from collections.abc import Iterable
-from meer21cm.util import check_unit_equiv
+from healpy.rotator import Rotator
 
 default_data_dir = meer21cm.__file__.rsplit("/", 1)[0] + "/data/"
 
@@ -14,142 +11,135 @@ default_data_dir = meer21cm.__file__.rsplit("/", 1)[0] + "/data/"
 class ForegroundSimulation:
     def __init__(
         self,
-        hp_nside=512,
-        verbose=False,
-        map_unit=units.K,
-        sync_map_file=None,
-        sync_indx_file=None,
-        sync_uni_indx=-2.55,
-        do_point_souce=False,
+        hp_nside=256,
         wproj=None,
         num_pix_x=None,
         num_pix_y=None,
-        sigma_beam_ch=None,
-        nu=None,
-        cache=False,
-        **fg_params,
+        backend="haslam",
+        pysm_preset_strings=["d1", "s1", "f1", "a1", "c1"],
+        sp_indx_for_haslam_backend=-2.0,
+        coord_system="C",
     ):
-        self.cache = cache
+        """
+        Foreground simulation class.
+        All outputs are in units of K_RJ (Kelvin).
+
+        Parameters
+        ----------
+        hp_nside: int
+            HEALPix nside.
+        wproj: tuple
+            WCS projection parameters.
+        num_pix_x: int
+            Number of pixels in x direction.
+        num_pix_y: int
+            Number of pixels in y direction.
+        backend: str, default 'haslam'
+            Backend to use for foreground simulation.
+            Options: 'gdsm' (Global Sky Model), 'pysm' (PySM3), 'haslam' (Haslam 408 MHz map).
+        pysm_preset_strings: list, default ['d1', 's1', 'f1', 'a1', 'c1']
+            List of PySM3 preset strings for included components.
+            Default: ['d1', 's1', 'f1', 'a1', 'c1'].
+        sp_indx_for_haslam_backend: float, default -2.0
+            Index of the spectral index for the Haslam 408 MHz map.
+            Only used for 'haslam' backend.
+        coord_system: str, default 'C'
+            Coordinate system to use for the foreground simulation.
+            Options: 'G' (Galactic), 'C' (Celestial), 'E' (Ecliptic).
+        """
+        self.backend = backend
+        assert backend in [
+            "gdsm",
+            "pysm",
+            "haslam",
+        ], "backend must be either 'gdsm', 'pysm' or 'haslam'"
         self.hp_nside = hp_nside
-        self.verbose = verbose
-        self.map_unit = map_unit
         self.wproj = wproj
         self.num_pix_x = num_pix_x
         self.num_pix_y = num_pix_y
-        self.sigma_beam_ch = sigma_beam_ch
-        self.nu = nu
-        if not check_unit_equiv(map_unit, units.K):
-            if not check_unit_equiv(map_unit, units.Jy):
-                raise (
-                    ValueError,
-                    "map unit has be to either temperature or flux density.",
-                )
-            else:
-                self.map_unit_type = "F"
-        else:
-            self.map_unit_type = "T"
-        self.sync_map_file = sync_map_file
-        self.sync_indx_file = sync_indx_file
-        self.sync_nside = None
-        self.sync_unit = None
-        self.sync_freq = None
-        self.sync_uni_indx = sync_uni_indx
-        self.do_point_souce = do_point_souce
-        self.__dict__.update(fg_params)
+        assert hp.isnsideok(hp_nside), "hp_nside must be a valid HEALPix nside"
+        self.pysm_preset_strings = pysm_preset_strings
+        self.sp_indx_for_haslam_backend = sp_indx_for_haslam_backend
+        self.coord_system = coord_system
 
-    @property
-    def sync_map(self):
-        file = self.sync_map_file
-        if file is None:
-            file = default_data_dir + "haslam408_dsds_Remazeilles2014.fits"
-        sync_map, self.sync_nside, self.sync_unit, self.sync_freq = read_healpix_fits(
-            file
+    def healpix_gen_haslam(self, freq):
+        """
+        Generate HEALPix map of foregrounds using Haslam 408 MHz map.
+        """
+        sp_indx = self.sp_indx_for_haslam_backend
+        freq = np.atleast_1d(freq)
+        haslam_map, hp_nside, map_unit, map_freq = read_healpix_fits(
+            default_data_dir + "haslam408_dsds_Remazeilles2014.fits"
         )
-        if check_unit_equiv(self.sync_unit, units.Jy):
-            if self.verbose:
-                print("convert " + file + " to temperature")
-            sync_map = (sync_map * self.sync_unit / 1 / units.Jy).to("").value
-            sync_map = convert_hpmap_in_jy_to_temp(sync_map, self.sync_freq)
-        return sync_map
+        assert check_unit_equiv(map_unit, units.K), "map unit must be temperature"
+        haslam_map = (haslam_map * map_unit).to(units.K).value
+        haslam_map = hp.ud_grade(haslam_map, self.hp_nside)
+        r = Rotator(coord=["G", self.coord_system])
+        haslam_map = r.rotate_map_pixel(haslam_map)
+        cube = haslam_map[None, :] * ((freq / map_freq) ** sp_indx)[:, None]
+        return cube
 
-    @property
-    def sync_spindx(self):
-        sync_indx_file = self.sync_indx_file
-        if sync_indx_file is None:
-            use_map = False
-            uni_indx = self.sync_uni_indx
-        elif isinstance(sync_indx_file, float):
-            uni_indx = sync_indx_file
-            use_map = False
-        elif isinstance(sync_indx_file, Iterable):
-            if isinstance(sync_indx_file[0], str) and isinstance(
-                sync_indx_file[1], str
-            ):
-                map1 = sync_indx_file[0]
-                map2 = sync_indx_file[1]
-                use_map = True
-        else:
-            raise TypeError("cannot recognise sync_indx_file")
-        if use_map:
-            map1_val, map1_nside, map1_unit, map1_freq = read_healpix_fits(map1)
-            map2_val, map2_nside, map2_unit, map2_freq = read_healpix_fits(map2)
-            if map1_nside != self.sync_nside:
-                map1_val = hp.ud_grade(map1_val, self.sync_nside)
-            if map2_nside != self.sync_nside:
-                map2_val = hp.ud_grade(map2_val, self.sync_nside)
-            unit_conv = (1 * map1_unit / map2_unit).to("").value
-            sp_indx = np.log(unit_conv * map1_val / map2_val) / np.log(
-                map1_freq / map2_freq
-            )
-        else:
-            sp_indx = np.ones(hp.nside2npix(self.sync_nside)) * uni_indx
-        return sp_indx
+    def healpix_gen_gdsm(self, freq):
+        """
+        Generate HEALPix map of foregrounds using Global Sky Model.
 
-    @property
-    def sync_cube(self):
-        if self.nu is None:
-            return None
-        sync_cube = self.fg_cube(self.nu, source_type="sync")
-        if self.cache:
-            self.sync_cache = sync_cube
-        return sync_cube
+        Parameters
+        ----------
+        freq: float or array-likesss
+            Frequency in Hz.
+        """
+        from pygdsm import GlobalSkyModel16
 
-    def fg_cube(
+        gsm = GlobalSkyModel16(
+            freq_unit="MHz",
+            data_unit="TRJ",
+            resolution="hi" if self.hp_nside > 64 else "low",
+        )
+        cube = gsm.generate(freq / 1e6)
+        cube = hp.ud_grade(cube, self.hp_nside)
+        r = Rotator(coord=["G", self.coord_system])
+        if len(freq) == 1:
+            cube = cube[None, :]
+        for i in range(cube.shape[0]):
+            cube[i] = r.rotate_map_pixel(cube[i])
+        return cube
+
+    def healpix_gen_pysm(self, freq):
+        """
+        Generate HEALPix map of foregrounds using PySM3.
+        """
+        import pysm3
+
+        pysm = pysm3.Sky(nside=self.hp_nside, preset_strings=self.pysm_preset_strings)
+        cube = []
+        for f in freq:
+            cube_i = pysm.get_emission(f * units.Hz)[0].value / 1e6
+            cube_i = hp.ud_grade(cube_i, self.hp_nside)
+            r = Rotator(coord=["G", self.coord_system])
+            cube_i = r.rotate_map_pixel(cube_i)
+            cube.append(cube_i)
+        cube = np.array(cube)
+        return cube
+
+    def fg_wcs_cube(
         self,
         freq,
-        source_type="sync",
     ):
+        freq = np.atleast_1d(freq)
         wproj = self.wproj
         xdim = self.num_pix_x
         ydim = self.num_pix_y
-        sigma_beam_ch = self.sigma_beam_ch
-        if source_type == "sync":
-            in_map = self.sync_map
-            spindx = self.sync_spindx
-            base_freq = self.sync_freq
-        freq = np.array([freq]).ravel()
-        if hp.get_nside(in_map) != self.hp_nside:
-            in_map = hp.ud_grade(in_map, self.hp_nside)
-            spindx = hp.ud_grade(spindx, self.hp_nside)
-        in_map = in_map[:, None]
-        spindx = spindx[:, None]
-        freq = freq[None, :]
-        out_map = in_map * (freq / base_freq) ** spindx
-        if sigma_beam_ch is not None:
-            for ch_id in range(out_map.shape[-1]):
-                out_map[:, ch_id] = hp.smoothing(
-                    out_map[:, ch_id], sigma=sigma_beam_ch[ch_id] * np.pi / 180
-                )
+        out_map = getattr(self, f"healpix_gen_{self.backend}")(freq)
         if wproj is not None:
             xx, yy = np.meshgrid(
                 np.arange(xdim),
                 np.arange(ydim),
                 indexing="ij",
             )
-            out_map_proj = np.zeros(xx.shape + (out_map.shape[-1],))
-            for ch_id in range(out_map.shape[-1]):
+            out_map_proj = np.zeros(xx.shape + (out_map.shape[0],))
+            for ch_id in range(out_map.shape[0]):
                 out_map_proj[:, :, ch_id] = healpix_to_wcs(
-                    out_map[:, ch_id], xx, yy, wproj
+                    out_map[ch_id], xx, yy, wproj
                 )
             out_map = out_map_proj
         return out_map