+++ tests/test_dataanalysis.py
@@ -4,7 +4,7 @@ import numpy as np
 from astropy import units, constants
 import pytest
 from astropy.wcs.utils import proj_plane_pixel_area
-from meer21cm.util import freq_to_redshift, center_to_edges, f_21, create_wcs_with_range
+from meer21cm.util import freq_to_redshift, center_to_edges, f_21, pca_clean
 from meer21cm.telescope import dish_beam_sigma
 
 
@@ -274,3 +274,71 @@ def test_trim_gal():
     sp._dec_gal = np.array([0, -2, 0])
     sp._z_gal = np.array([0.42, 0.42, 0.42])
     assert np.allclose(sp.trim_gal_to_range(), [1, 0, 0])
+
+
+def test_get_jackknife_patches():
+    ps = Specification(
+        survey="meerklass_2021",
+        band="L",
+    )
+    with pytest.raises(AssertionError):
+        ps.get_jackknife_patches(
+            ra_patch_num=4,
+            dec_patch_num=4,
+            nu_patch_num=2,
+        )
+    ps.ra_range = (334, 357)
+    ps.dec_range = (-35, -26.5)
+    mask_arr = ps.get_jackknife_patches(ra_patch_num=8, dec_patch_num=4, nu_patch_num=2)
+    assert mask_arr.shape == (8, 4, 2, ps.num_pix_x, ps.num_pix_y, ps.nu.size)
+    # check ra dec makes sense
+    mask_pixel = mask_arr.sum((2, 3, 4, 5))
+    mask_pixel_mean = mask_pixel.mean()
+    mask_pixel_std = mask_pixel.std()
+    assert mask_pixel_std / mask_pixel_mean < 5e-2
+    # check nu makes sense
+    nu_pixel = mask_arr.sum((0, 1, 2, 3, 4))
+    assert nu_pixel.std() / nu_pixel.mean() < 1e-2
+
+
+def test_create_white_noise_map():
+    ra_range_MK = (334, 357)
+    dec_range_MK = (-35, -26.5)
+    ps = Specification(
+        band="L",  # band and survey will produce some pre-defined cuts to select
+        survey="meerklass_2021",  # the clean frequency sub-band
+        ra_range=ra_range_MK,
+        dec_range=dec_range_MK,
+    )
+    noise_map = ps.create_white_noise_map(
+        0.1,
+    )
+    std = ((noise_map * np.sqrt(ps.counts))[ps.counts > 0]).std()
+    assert np.allclose(std, 0.1, rtol=5e-3)
+    ps.counts = np.random.uniform(1, 100, size=ps.data.shape) * ps.W_HI
+    noise_map = ps.create_white_noise_map(0.1, counts=ps.counts)
+    std = ((noise_map * np.sqrt(ps.counts))[ps.counts > 0]).std()
+    assert np.allclose(std, 0.1, rtol=5e-3)
+
+
+def test_check_is_map_noiselike_using_pca():
+    ra_range_MK = (334, 357)
+    dec_range_MK = (-35, -26.5)
+    ps = Specification(
+        band="L",  # band and survey will produce some pre-defined cuts to select
+        survey="meerklass_2021",  # the clean frequency sub-band
+        ra_range=ra_range_MK,
+        dec_range=dec_range_MK,
+    )
+    noise_map = ps.create_white_noise_map(
+        0.1,
+    )
+    N_fg = 15
+    res_map, A_mat = pca_clean(noise_map, N_fg, weights=ps.W_HI, return_A=True)
+    ps.data = noise_map
+    res_var, noise_var = ps.check_is_map_noiselike_using_pca(A_mat, sigma_N=0.1)
+    assert np.allclose(
+        res_var,
+        noise_var,
+        rtol=1e-1,
+    )