+++ tests/test_mock.py
@@ -424,14 +424,14 @@ def test_flat_sky():
     mock.include_sky_sampling = [False, False]
     mock.compensate = False
     ratio = mock.auto_power_3d_1 / mock.auto_power_tracer_1_model
-    assert np.abs(ratio.mean() - 1) < 1e-1
+    assert np.abs(ratio.mean() - 1) < 2e-1
     mock.propagate_mock_tracer_to_gal_cat()
     mock.grid_gal_to_field()
     mock.weights_2 = np.ones_like(mock.field_2)
     mock.compensate = False
     shot_noise = np.prod(mock.box_len) / mock.field_2.sum()
     ratio = (mock.auto_power_3d_2 - shot_noise) / mock.auto_power_tracer_2_model
-    assert np.abs(ratio.mean() - 1) < 1e-1
+    assert np.abs(ratio.mean() - 1) < 2e-1
 
 
 def test_dndz():