+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ dev
 Features
 ++++++++
 * add get_shot_noise_galaxy function
+* add jackknife patch split and plot function
+* improve foreground simulation routine
 
 Fixes
 +++++